Filter AVS title out of page header
